backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
requests
pydantic
faster-whisper